Parse pGlyco Structures

Parse pGlyco-style structure characters into glycan graphs. See example below for the structure format.

Usage

parse_pglyco_struc(x, mode = "ne")

Arguments

x

A character vector of pGlyco-style structure strings.

mode

A character string, either "ne" or "dn". Default to "ne". "ne" is recommended in this case, for pGlyco-style structures don't have linkage information. For more information about "ne" and "dn", see glyrepr::as_glycan_graph().

Value

A glycan graph if x is a single character, or a list of glycan graphs if x is a character vector.

Examples

glycan <- parse_pglyco_struc("(N(F)(N(H(H(N))(H(N(H))))))")
print(glycan, verbose = TRUE)
#> Glycan Graph (NE)
#> F: 1, H: 4, N: 4
#> ------------------
#> N
#> ├─F
#> └─N
#>   └─H
#>     ├─H
#>     │ └─N
#>     └─H
#>       └─N
#>         └─H