Wrapper for Structure Parsers

This wrapper function adds the following functionalities to a structure parser function:

Check input type;

Vectorized parsing;

More informative error messages.

This wrapper is used by all parser interfaces in the package, e.g. parse_pglyco_struc(). If you are writing a new parser, you can use this function to wrap your parser function.

Usage

struc_parser_wrapper(x, parser, mode = "ne")

Arguments

x

A character vector of structure strings.

parser

A function that parses a single structure string and returns a glycan graph (see glyrepr::as_glycan_graph()).

mode

A character string, either "ne" or "dn".

Value

A glycan graph if x is a single character, or a list of glycan graphs if x is a character vector.